Fix compile warning in x86

+++ polyhook2/Instruction.hpp
@@ -194,7 +194,7 @@ public:
 		m_isRelative = true;
 		m_hasDisplacement = true;
 
-		assert(m_dispOffset + m_dispSize <= m_bytes.size() && m_dispSize <= sizeof(m_displacement.Relative));
+		assert(static_cast<size_t>(m_dispOffset) + m_dispSize <= m_bytes.size() && m_dispSize <= sizeof(m_displacement.Relative));
 		std::memcpy(&m_bytes[getDisplacementOffset()], &m_displacement.Relative, m_dispSize);
 	}
